move ModuleTree out of the test file
